backend/requirements.txt
dj-database-url==0.4.1
Django==1.10.1
djangorestframework==3.4.7
gunicorn==19.6.0
oauthlib==2.0.0
python-twitter==3.1
psycopg2==2.6.2
watson-developer-cloud==0.20.0
